fix(directory): wrong names & paths & organized
