+++ app/views/auth/auth.ejs
@@ -73,7 +73,7 @@
     <main class="flex-grow flex items-center justify-center p-8">
         <div class="py-8 px-6 md:py-12 md:px-8 max-w-md w-full">
             <div class="flex justify-center mb-8">
-                <img src="/assets/logo-text.png" class="h-12 hover:scale-90 transition cursor-pointer" alt="Logo">
+                <img src="/assets/images/logo-text.png" class="h-12 hover:scale-90 transition cursor-pointer" alt="Logo">
             </div>
             <h2 class="text-center text-2xl tracking-tight font-semibold text-gray-800"><%= __('auth.login.welcome') %></h2>
             <p class="text-gray-600 text-center mt-1 text-sm"><%= __('auth.login.subtitle') %></p>
