+++ app/views/auth/register.ejs
@@ -99,7 +99,7 @@
     <main class="flex-grow flex items-center justify-center p-8">
         <div class="py-8 px-6 md:py-12 md:px-8 max-w-md w-full">
             <div class="flex flex-col md:flex-row justify-between gap-y-3 items-center mb-8">
-                <img src="/assets/logo-text.png" class="h-12">
+                <img src="/assets/images/logo-text.png" class="h-12">
                 <div class="flex gap-x-2">
                     <% availableLocales.forEach(l => { %>
                         <a href="?lang=<%= l.code %>" class="text-xl hover:scale-110 transition-transform <%= locale === l.code ? 'grayscale-0' : 'grayscale opacity-50 hover:opacity-100' %>" title="<%= l.name %>">
