+++ app/views/auth/resetpassword.ejs
@@ -61,7 +61,7 @@
     <main id="mainContent" class="flex-grow flex items-center justify-center p-8">
         <div class="py-8 px-6 md:py-12 md:px-8 max-w-md w-full">
             <div class="flex justify-center mb-8">
-                <img src="/assets/logo-text.png" class="h-12 hover:scale-90 transition cursor-pointer" alt="Logo">
+                <img src="/assets/images/logo-text.png" class="h-12 hover:scale-90 transition cursor-pointer" alt="Logo">
             </div>
             <h2 class="text-center text-2xl tracking-tight font-semibold text-gray-800" style="font-family: 'Inter'"><%= __('auth.reset.heading') %></h2>
             <p class="text-gray-400 text-center mt-1 font-normal text-sm"><%= __('auth.reset.subtitle') %></p>
