+++ app/views/components/bottom.ejs
@@ -119,7 +119,7 @@
     overlay.className = 'page-loading-overlay';
     overlay.innerHTML = `
         <div class="flex justify-center">
-            <img src="/assets/logo-text.png" class="h-12 w-auto invert">
+            <img src="/assets/images/logo-text.png" class="h-12 w-auto invert">
         </div>
         <svg class="animate-spin h-5 w-5 text-white" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
             <circle class="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" stroke-width="4"></circle>
