+++ app/views/components/top.ejs
@@ -29,7 +29,7 @@
     <title><%= settings.name %></title>
 
     <link rel="stylesheet" href="/assets/tailwind.css">
-    <link rel="icon" type="image/x-icon" href="/assets/favicon.ico">
+    <link rel="icon" type="image/x-icon" href="/assets/images/favicon.ico">
 
     <script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js"></script>
 
@@ -38,7 +38,7 @@
     
     <link 
         rel="preload" 
-        href="/assets/Ndot-55.otf" 
+        href="/assets/others/Ndot-55.otf" 
         as="font" 
         type="font/otf" 
         crossorigin="anonymous"
@@ -139,7 +139,7 @@
 
         @font-face {
             font-family: 'Ndot-55';
-            src: url('/assets/Ndot-55.otf') format('opentype');
+            src: url('/assets/others/Ndot-55.otf') format('opentype');
             font-weight: normal;
             font-display: swap;
             font-style: normal;
@@ -230,7 +230,7 @@
         <!-- Left side -->
         <div class="items-center hover:bg-white/5 rounded-md px-1 md:px-3 py-1.5 transition duration-200 hover:scale-95 cursor-pointer"
             onclick="window.location.href='/dashboard'">
-            <img src="/assets/console.png" alt="<%= settings.name %> Logo" class="h-8 object-fill">
+            <img src="/assets/images/console.png" alt="<%= settings.name %> Logo" class="h-8 object-fill">
         </div>
 
         <!-- Right side -->
