+++ app/views/create.ejs
@@ -1,4 +1,4 @@
-<%- include('../components/top') %>
+<%- include('components/top') %>
 <!-- Main Content -->
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/AlertifyJS/1.13.1/css/alertify.css" integrity="sha512-MpdEaY2YQ3EokN6lCD6bnWMl5Gwk7RjBbpKLovlrH6X+DRokrPRAF3zQJl1hZUiLXfo2e9MrOt+udOnHCAmi5w==" crossorigin="anonymous" referrerpolicy="no-referrer" />
 <main class="flex-grow container mx-auto px-4 py-8">
@@ -791,4 +791,4 @@ alertify.set('notifier','position', 'top-right');
     }
 </style>
 </main>
-<%- include('../components/bottom') %>
+<%- include('components/bottom') %>
