+++ app/views/error/404.ejs
@@ -11,7 +11,7 @@
     <main class="flex-grow flex items-center justify-center p-8">
         <div class="py-8 px-6 md:py-12 md:px-8 max-w-xl w-full">
             <div class="flex justify-center mb-8">
-                <img src="/assets/logo-text.png" class="h-12" alt="Logo">
+                <img src="/assets/images/logo-text.png" class="h-12" alt="Logo">
             </div>
             <h1 class="text-center text-4xl tracking-tight font-bold text-gray-900 mb-4"><%= __('errors.404.title') %></h1>
             <h2 class="text-center text-2xl tracking-tight font-semibold text-gray-800 mb-2"><%= __('errors.404.heading') %></h2>
