+++ app/views/pages.json
@@ -12,9 +12,9 @@
         "dashboard": "dashboard.ejs",
         "auth": "auth/auth.ejs",
         "auth/register": "auth/register.ejs",
-        "auth/reset-password": "auth/passwordreset.ejs",
+        "auth/reset-password": "auth/resetpassword.ejs",
         "servers": "servers.ejs",
-        "server/new": "server/create.ejs",
+        "server/new": "create.ejs",
         "server/edit": "server/edit.ejs",
         "server/manage": "server/manage.ejs",
         "server/allocations": "server/allocations.ejs",
