+++ app/views/website/components/footer.ejs
@@ -8,7 +8,7 @@
         <div class="mx-auto max-w-7xl px-6 pb-8 pt-16 sm:pt-24 lg:px-8 lg:pt-32">
             <div class="xl:grid xl:grid-cols-3 xl:gap-8">
                 <div class="space-y-8">
-                    <img class="h-12 invert" src="/assets/logo-text.png" alt="<%= settings.name %> Logo">
+                    <img class="h-12 invert" src="/assets/images/logo-text.png" alt="<%= settings.name %> Logo">
                     <p class="text-sm leading-6 text-gray-300">
                         <% if (locationCount === 1) { %>
                             <%= __('website.footer.about', { name: settings.name, scope: __('website.footer.scope_one', { country: firstLocation.country }) }) %>
