+++ app/views/website/components/header.ejs
@@ -5,7 +5,7 @@
             <div class="flex items-center justify-between">
                 <div class="flex lg:flex-1">
                     <a href="/" class="flex items-center">
-                        <img src="/assets/logo-text.png" class="h-10" alt="<%= settings.name %> Logo">
+                        <img src="/assets/images/logo-text.png" class="h-10" alt="<%= settings.name %> Logo">
                     </a>
                 </div>
