+++ app/views/website/vps.ejs
@@ -15,7 +15,7 @@ const firstLocation = locationCount === 1 ? Object.values(settings.api.client.lo
         body { font-family: 'Inter', sans-serif; }
         @font-face {
             font-family: 'Ndot-55';
-            src: url('/assets/Ndot-55.otf') format('opentype');
+            src: url('/assets/others/Ndot-55.otf') format('opentype');
             font-weight: normal;
             font-display: swap;
         }
