+++ app/views/website/website.ejs
@@ -20,7 +20,7 @@ const firstLocation = locationCount === 1 ? Object.values(settings.api.client.lo
         }
         @font-face {
             font-family: 'Ndot-55';
-            src: url('/assets/Ndot-55.otf') format('opentype');
+            src: url('/assets/others/Ndot-55.otf') format('opentype');
             font-weight: normal;
             font-display: swap;
         }
